Use enums rather than basic types for compact unwind row
Or, use the type system :)

Test Plan
=========

CI + some simple benchmarks with `--release` to ensure that performance
did not degrade

```
$ hyperfine "target/lightswitch-new --show-info /proc
/2154884/root/nix/store/dwzq3jgydz8wwa7xh86f8yfxg4isnjkj-clickhouse-24.3.2.23/bin/clickhouse" "target/light
switch-main --show-info /proc/2154884/root/nix/store/dwzq3jgydz8wwa7xh86f8yfxg4isnjkj-clickhouse-24.3.2.23/
bin/clickhouse"
Benchmark 1: target/lightswitch-new --show-info /proc/2154884/root/nix/store/dwzq3jgydz8wwa7xh86f8yfxg4isnjkj-clickhouse-24.3.2.23/bin/clickhouse
  Time (mean ± σ):     805.8 ms ±   9.5 ms    [User: 778.6 ms, System: 24.9 ms]
  Range (min … max):   792.8 ms … 817.7 ms    10 runs

Benchmark 2: target/lightswitch-main --show-info /proc/2154884/root/nix/store/dwzq3jgydz8wwa7xh86f8yfxg4isnjkj-clickhouse-24.3.2.23/bin/clickhouse
  Time (mean ± σ):     804.6 ms ±   9.9 ms    [User: 777.9 ms, System: 24.8 ms]
  Range (min … max):   792.7 ms … 825.0 ms    10 runs

Summary
  target/lightswitch-main --show-info /proc/2154884/root/nix/store/dwzq3jgydz8wwa7xh86f8yfxg4isnjkj-clickhouse-24.3.2.23/bin/clickhouse ran
    1.00 ± 0.02 times faster than target/lightswitch-new --show-info /proc/2154884/root/nix/store/dwzq3jgydz8wwa7xh86f8yfxg4isnjkj-clickhouse-24.3.2.23/bin/clickhouse
```

+++ src/bpf/profiler_bindings.rs
@@ -82,8 +82,8 @@ impl From<&CompactUnwindRow> for stack_unwind_row_t {
             // https://github.com/rust-lang/rust-bindgen/issues/923#issuecomment-2385554573
             pc_low: (row.pc & LOW_PC_MASK as u64) as u16,
             cfa_offset: row.cfa_offset,
-            cfa_type: row.cfa_type,
-            rbp_type: row.rbp_type,
+            cfa_type: row.cfa_type as u8,
+            rbp_type: row.rbp_type as u8,
             rbp_offset: row.rbp_offset,
         }
     }
