+++ src/cli/main.rs
@@ -226,7 +226,7 @@ fn show_unwind_info(path: &str) {
         let rbp_offset = compact_row.rbp_offset;
         println!(
             "pc: {:x} cfa_type: {:<2} rbp_type: {:<2} cfa_offset: {:<4} rbp_offset: {:<4}",
-            pc, cfa_type, rbp_type, cfa_offset, rbp_offset
+            pc, cfa_type as u8, rbp_type as u8, cfa_offset, rbp_offset
         );
     }
 }
